modify playGame function main loop
Previously the function would loop for 5 times, playing 5 rounds and
then, if there was a tie, it would keep calling the playRound function
until there was a winner.

Now the function keeps calling playRound until either the computer or
the player reaches a score of 3. It essentially implements a
best of 5/first to 3 scoring system and stops executing if either the
player or computer reaches a score of 3 before getting to 5 rounds.