modify playGame function main loop
The first implementation was only "best of 5", this meant that the
player would necessarily need to play at least 5 rounds, even if the
player or computer reached a score of 3 before 5 rounds.

To fix this, the second implementation was "first to 3", whoever
reached a score of 3 first would win. This created another issue where
if neither the player nor the computer reached a score of 3 in 5 rounds
the function playGame would continue calling playRound until a score of
3 was reached, despite the fact that there was already a clear winner.

Now the function's main loop combines both of these scoring systems to
truly implement a best of 5 scoring system that also ends the game if
the player or computer reaches a score of 3 before these 5 rounds.

Examples:

L - Player loses
W - Player wins
T - Player and computer both play the same choice in a given rounds

L W L T T

The computer wins 2 rounds, the player wins 1 round and 2 of the rounds
are a tie. In this case, the computer wins the game. In the previous
"first to 3" implementation, the game would keep on going until either
the player or the computer reached a score of 3.

W W T W

The player wins the game and the function playGame stops execution, no
need to play another round as the player already won 3 rounds. The
"best of 5" only implementation would continue to play another round.