update final score design
Now the text for the final score is centralized, also the method for
announcing the final score just updates the text content of the score
paragraph instead of appending a new paragraph.
This means that if the player plays more than 1 game the score will
just be updated with each game instead of appended below the previous
score.